feat(auth): model and entity added
