+++ internal/domain/entities/auth_user.go
@@ -1,0 +1,12 @@
+package entities
+
+import "gorm.io/gorm"
+
+type User struct {
+	gorm.Model
+
+	FullName       string `json:"full_name"`
+	Specialization string `json:"specialization"`
+	Login          string `json:"login"`
+	PasswordHash   string `json:"-"`
+}
